+++ src/Dungeon.cpp
@@ -46,18 +46,18 @@ Dungeon::Dungeon(int argc, char* argv[]) : textLines{new std::string, new std::s
     if (this->settings->doLoad()) {
         load_from_file(this);
     } else {
-        auto stairCount = u_short(Dice::RandomNumberBetween(FLOOR_STAIRS_MIN, FLOOR_STAIRS_MAX));
+        u_short stairCount = Dice::RandomNumberBetween(FLOOR_STAIRS_MIN, FLOOR_STAIRS_MAX);
 
         for (index = 0; index < Dice::RandomNumberBetween(DUNGEON_FLOORS_MIN, DUNGEON_FLOORS_MAX); index++) {
             this->floors.push_back(
                     new Floor(
                             this,
                             index,
-                            u_short(Dice::RandomNumberBetween(FLOOR_ROOMS_MIN, FLOOR_ROOMS_MAX)),
+                            Dice::RandomNumberBetween(FLOOR_ROOMS_MIN, FLOOR_ROOMS_MAX),
                             stairCount,
                             stairCount,
-                            u_short(Dice::RandomNumberBetween(FLOOR_MONSTERS_MIN, FLOOR_MONSTERS_MAX)),
-                            u_short(Dice::RandomNumberBetween(FLOOR_OBJECTS_MIN, FLOOR_OBJECTS_MAX))
+                            Dice::RandomNumberBetween(FLOOR_MONSTERS_MIN, FLOOR_MONSTERS_MAX),
+                            Dice::RandomNumberBetween(FLOOR_OBJECTS_MIN, FLOOR_OBJECTS_MAX)
                     )
             );
         }
@@ -134,41 +134,26 @@ std::string* Dungeon::appendText(const std::string& text) {
 }
 
 std::string* Dungeon::prependText(const std::string* format, ...) {
-    u_char index;
     char buffer[DUNGEON_FLOOR_WIDTH + 1];
+    
     va_list args;
     va_start(args, format);
-
-    for (index = DUNGEON_TEXT_LINES - 1; index > 0; index--) {
-        this->textLines[index]->assign(*(this->textLines[index - 1]));
-    }
-
     vsnprintf(buffer, DUNGEON_FLOOR_WIDTH + 1, format->c_str(), args);
-    this->textLines[0]->assign(buffer, DUNGEON_FLOOR_WIDTH);
-
     va_end(args);
-
-    return this->textLines[0];
+    
+    
+    return this->prependText(buffer);
 }
 
 std::string* Dungeon::appendText(const std::string* format, ...) {
-    u_char index;
     char buffer[DUNGEON_FLOOR_WIDTH + 1];
-
-    for (index = DUNGEON_TEXT_LINES - 1; index > 0; index--) {
-        this->textLines[index - 1]->assign(*(this->textLines[index]));
-    }
 
     va_list args;
     va_start(args, format);
-
-    vsnprintf(buffer, DUNGEON_FLOOR_WIDTH, format->c_str(), args);
-
+    vsnprintf(buffer, DUNGEON_FLOOR_WIDTH + 1, format->c_str(), args);
     va_end(args);
-
-    this->textLines[DUNGEON_TEXT_LINES - 1]->assign(buffer, DUNGEON_FLOOR_WIDTH);
-
-    return this->textLines[DUNGEON_TEXT_LINES - 1];
+    
+    return this->appendText(buffer);
 }
 
 MonsterTemplate* Dungeon::randomMonsterTemplate() {
@@ -331,12 +316,6 @@ Dungeon* Dungeon::setPlayer(Player* player) {
 
 Dungeon* Dungeon::setBoss(Monster* boss) {
     this->boss = boss;
-
-    return this;
-}
-
-Dungeon* Dungeon::setWindow(WINDOW* window) {
-    this->window = window;
 
     return this;
 }