+++ src/terrains/border.h
@@ -10,7 +10,7 @@
 namespace App {
     class Border : public Terrain {
     public:
-        Border(Floor* floor, u_short id, u_char x, u_char y, u_char character);
+        Border(Floor* floor, u_char x, u_char y, u_char character);
         ~Border();
 
         /** GETTERS **/