+++ src/terrains/corridor.cpp
@@ -1,6 +1,6 @@
 #include "corridor.h"
 
-Corridor::Corridor(Floor* floor, u_short id, u_char x, u_char y) : Terrain(floor, id, x, y) {
+Corridor::Corridor(Floor* floor, u_char x, u_char y) : Terrain(floor, x, y) {
     this->character = CORRIDOR_CHARACTER;
     this->hardness = CORRIDOR_HARDNESS;
     this->type = TERRAIN_CORRIDOR;