+++ src/terrains/rock.cpp
@@ -1,6 +1,6 @@
 #include "rock.h"
 
-Rock::Rock(Floor* floor, u_short id, u_char x, u_char y, u_char hardness) : Terrain(floor, id, x, y) {
+Rock::Rock(Floor* floor, u_char x, u_char y, u_char hardness) : Terrain(floor, x, y) {
     this->character = ROCK_CHARACTER;
     this->hardness = hardness;
     this->type = TERRAIN_ROCK;