+++ src/terrains/rock.h
@@ -10,7 +10,7 @@
 namespace App {
     class Rock : public Terrain {
     public:
-        Rock(Floor* floor, u_short id, u_char x, u_char y, u_char hardness);
+        Rock(Floor* floor, u_char x, u_char y, u_char hardness);
         ~Rock();
 
         /** GETTERS **/